OpenAI Whisper/requirements.txt
openai-whisper
torch
torchaudio
soundfile
numpy
sounddevice
webrtcvad
noisereduce